requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
sqlalchemy==2.0.48
pydantic==2.12.5
pydantic[email]==2.12.5
python-jose==3.5.0
passlib==1.7.4
bcrypt==4.0.1
cryptography==46.0.5
psycopg2-binary==2.9.11
python-multipart==0.0.22
email-validator==2.3.0
python-dotenv==1.1.0
